pyproject.toml
[project]
name = "env-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "firecrawl>=2.1.2",
    "gitpython>=3.1.44",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "pandas>=2.3.1",
]

[dependency-groups]
dev = ["pyright>=1.1.398", "pytest>=8.3.5", "ruff>=0.11.4"]
